order_from

+++ app/src/main/java/me/labs/corobox/corobox/view/main_screen/make_order_screen/MakeOrderActivityView.java
@@ -6,7 +6,6 @@ import android.app.TimePickerDialog;
 import android.os.Bundle;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -17,16 +16,9 @@ import android.widget.TextView;
 import android.widget.TimePicker;
 import android.widget.Toast;
 
-import com.google.gson.ExclusionStrategy;
-import com.google.gson.FieldAttributes;
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-
 import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.Calendar;
-import java.util.Date;
-import java.util.GregorianCalendar;
+import java.util.List;
 import java.util.Locale;
 import java.util.UUID;
 
@@ -35,22 +27,21 @@ import javax.inject.Inject;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import io.realm.Realm;
-import io.realm.RealmObject;
+import io.realm.RealmList;
 import me.labs.corobox.corobox.R;
 import me.labs.corobox.corobox.app.CoroboxApp;
 import me.labs.corobox.corobox.common.BaseActivity;
 import me.labs.corobox.corobox.common.FragmentType;
-import me.labs.corobox.corobox.common.adapters.CategoriesOrderRealmAdapter;
-import me.labs.corobox.corobox.common.serializers.AddressSerializer;
-import me.labs.corobox.corobox.common.serializers.CategoryNumberModelSerializer;
-import me.labs.corobox.corobox.common.serializers.OrderModelToSerializer;
+import me.labs.corobox.corobox.common.adapters.CategoriesOrderFromRealmAdapter;
+import me.labs.corobox.corobox.common.adapters.CategoriesOrderToRealmAdapter;
 import me.labs.corobox.corobox.di.IHasComponent;
 import me.labs.corobox.corobox.di.components.ICoroboxAppComponent;
 import me.labs.corobox.corobox.di.components.activities.DaggerIMakeOrderActivityComponent;
 import me.labs.corobox.corobox.di.components.activities.IMakeOrderActivityComponent;
 import me.labs.corobox.corobox.di.modules.activities.MakeOrderActivityModule;
 import me.labs.corobox.corobox.model.realm.AddressModel;
-import me.labs.corobox.corobox.model.realm.CardModel;
+import me.labs.corobox.corobox.model.realm.Box;
+import me.labs.corobox.corobox.model.realm.OrderModelFrom;
 import me.labs.corobox.corobox.model.realm.OrderModelTo;
 import me.labs.corobox.corobox.presenter.make_order_screen.IMakeOrderActivityPresenter;
 
@@ -75,7 +66,11 @@ public class MakeOrderActivityView extends BaseActivity implements IMakeOrderAct
     private Realm realm;
 
     private String uuid;
-    private CategoriesOrderRealmAdapter categoriesOrderRealmAdapter;
+    private String type;
+    private CategoriesOrderToRealmAdapter categoriesOrderToRealmAdapter;
+    private CategoriesOrderFromRealmAdapter categoriesOrderFromRealmAdapter;
+
+    private OrderModelFrom orderModelFrom;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -88,6 +83,7 @@ public class MakeOrderActivityView extends BaseActivity implements IMakeOrderAct
         setTitle("Новый заказ");
 
         uuid = getIntent().getExtras().getString("uuid");
+        type = getIntent().getExtras().getString("type");
 
         initComponents();
         getWindow().setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_STATE_ALWAYS_HIDDEN);
@@ -142,11 +138,42 @@ public class MakeOrderActivityView extends BaseActivity implements IMakeOrderAct
             Toast.makeText(this, "Адрес по умолчанию не найден", Toast.LENGTH_SHORT).show();
         }
 
-        OrderModelTo orderModel = realm.where(OrderModelTo.class).equalTo("uuid_inner", uuid).findFirst();
+        if (type.equals("TO")) {
+            OrderModelTo orderModel = realm.where(OrderModelTo.class).equalTo("uuid_inner", uuid).findFirst();
+            categoriesOrderToRealmAdapter = new CategoriesOrderToRealmAdapter(orderModel.getCategoryNumberModel().createSnapshot(), presenter, false);
+            recyclerView.setAdapter(categoriesOrderToRealmAdapter);
+        } else if (type.equals("FROM")) {
+            OrderModelFrom orderModelFrom = generateOrderFrom();
+            categoriesOrderFromRealmAdapter = new CategoriesOrderFromRealmAdapter(orderModelFrom.getCategoryNumberModel(), presenter);
+            recyclerView.setAdapter(categoriesOrderFromRealmAdapter);
+        }
+    }
 
-        categoriesOrderRealmAdapter = new CategoriesOrderRealmAdapter(orderModel.getCategoryNumberModel().createSnapshot(), presenter, false);
+    private OrderModelFrom generateOrderFrom() {
+        orderModelFrom = new OrderModelFrom();
 
-        recyclerView.setAdapter(categoriesOrderRealmAdapter);
+        AddressModel address = realm.where(AddressModel.class).equalTo("useAsDefault", true).findFirst();
+        orderModelFrom.setAddressModel(address);
+
+        try {
+            orderModelFrom.setTill(dateToTimestamp(stringToDate(
+                    date.getText().toString().trim() + " " +
+                            time.getText().toString().trim(),
+                    "dd.MM.yyyy HH:mm", Locale.getDefault())));
+        } catch (ParseException e) {
+            e.printStackTrace();
+        }
+
+        RealmList<Box> boxes = new RealmList<>();
+        List<String> stuffUuid = getIntent().getExtras().getStringArrayList("list");
+        if (stuffUuid != null) {
+            for (String uuid : stuffUuid) {
+                Box box = realm.where(Box.class).equalTo("uuid", uuid).findFirst();
+                boxes.add(box);
+            }
+        }
+        orderModelFrom.setCategoryNumberModel(boxes);
+        return orderModelFrom;
     }
 
     @Override
@@ -168,30 +195,34 @@ public class MakeOrderActivityView extends BaseActivity implements IMakeOrderAct
         int id = item.getItemId();
 
         if (id == R.id.order) {
-            if (addressFlat.getText().toString().length() == 0) {
-                Toast.makeText(this, "Заполнены не все поля", Toast.LENGTH_SHORT).show();
-            } else {
-                realm.beginTransaction();
-                OrderModelTo orderModel = realm.where(OrderModelTo.class).equalTo("uuid_inner", uuid).findFirst();
-                orderModel.setUuid_inner(UUID.randomUUID().toString());
-                AddressModel addressModel = realm.where(AddressModel.class).equalTo("useAsDefault", true).findFirst();
-                orderModel.setOrderId(100000 + (5 + (int)(Math.random() * ((999999 - 100000) + 1))));
-                try {
-                    orderModel.setTill(dateToTimestamp(stringToDate(
-                            date.getText().toString().trim() + " " +
-                                    time.getText().toString().trim(),
-                            "dd.MM.yyyy HH:mm", Locale.getDefault())));
-                } catch (ParseException e) {
-                    e.printStackTrace();
-                }
-                orderModel.setAddressModel(addressModel);
-                orderModel.setType("FROM");
-                orderModel.setDate(date.getText().toString() + " " + time.getText().toString());
-                orderModel.setStatus("ORDERED");
-                realm.copyToRealm(orderModel);
-                realm.commitTransaction();
+            if (type.equals("TO")) {
+                if (addressFlat.getText().toString().length() == 0) {
+                    Toast.makeText(this, "Заполнены не все поля", Toast.LENGTH_SHORT).show();
+                } else {
+                    realm.beginTransaction();
+                    OrderModelTo orderModel = realm.where(OrderModelTo.class).equalTo("uuid_inner", uuid).findFirst();
+                    orderModel.setUuid_inner(UUID.randomUUID().toString());
+                    AddressModel addressModel = realm.where(AddressModel.class).equalTo("useAsDefault", true).findFirst();
+                    orderModel.setOrderId(100000 + (5 + (int) (Math.random() * ((999999 - 100000) + 1))));
+                    try {
+                        orderModel.setTill(dateToTimestamp(stringToDate(
+                                date.getText().toString().trim() + " " +
+                                        time.getText().toString().trim(),
+                                "dd.MM.yyyy HH:mm", Locale.getDefault())));
+                    } catch (ParseException e) {
+                        e.printStackTrace();
+                    }
+                    orderModel.setAddressModel(addressModel);
+                    orderModel.setType("FROM");
+                    orderModel.setDate(date.getText().toString() + " " + time.getText().toString());
+                    orderModel.setStatus("ORDERED");
+                    realm.copyToRealm(orderModel);
+                    realm.commitTransaction();
 
-                presenter.putOrder(orderModel);
+                    presenter.putOrder(orderModel);
+                }
+            } else if (type.equals("FROM")) {
+                presenter.putOrderFrom(orderModelFrom);
             }
             return true;
         }
@@ -211,7 +242,7 @@ public class MakeOrderActivityView extends BaseActivity implements IMakeOrderAct
 
     @Override
     public void updateList() {
-        categoriesOrderRealmAdapter.notifyDataSetChanged();
+        categoriesOrderToRealmAdapter.notifyDataSetChanged();
     }
 
     @Override
@@ -227,9 +258,11 @@ public class MakeOrderActivityView extends BaseActivity implements IMakeOrderAct
 
     @Override
     public void orderSuccessfullyAdded() {
-        realm.beginTransaction();
-        realm.where(OrderModelTo.class).equalTo("status", "ORDERED").findAll().deleteAllFromRealm();
-        realm.commitTransaction();
+        if (type.equals("TO")) {
+            realm.beginTransaction();
+            realm.where(OrderModelTo.class).equalTo("status", "ORDERED").findAll().deleteAllFromRealm();
+            realm.commitTransaction();
+        }
         Toast.makeText(this, "Заказ передан в обработку", Toast.LENGTH_SHORT).show();
         finish();
     }
